+++ src/views/wny/workflows/Workflows.vue
@@ -13,7 +13,6 @@
 
         <v-client-table :columns="columns" :data="data" :options="options" :theme="theme" id="dataTable">
           <p slot="actions" slot-scope="props">
-            <!--            <a :href="'#/roles/' + props.row.id" class="icon-eye action-icon"></a>-->
             <a :href="'#/admin/workflow/' + props.row.id + '/edit'" class="icon-pencil action-icon"></a>
             <a class="icon-trash" v-on:click="deleteWorkflow(props.row.id)" style="cursor: pointer"></a>
           </p>
@@ -74,17 +73,17 @@
     methods: {
       deleteWorkflow: function (id) {
         deleteWorkflow(id)
-            .then(() => this.stageDeletingSuccessful())
-            .catch((request) => this.stageDeletingFailed(request));
+            .then(() => this.workflowDeletingSuccessful())
+            .catch((request) => this.workflowDeletingFailed(request));
       },
-      stageDeletingSuccessful() {
+      workflowDeletingSuccessful() {
         this.errors = false;
         this.error = false;
         this.flash('The Workflow is deleted.', 'success', {timeout: VUE_APP_FLASH_TIMEOUT});
 
         this.downloadData();
       },
-      stageDeletingFailed(req) {
+      workflowDeletingFailed(req) {
         this.errors = false;
         this.error = 'The Workflow Deleting failed! ' + req;
         console.log(req);